@@ -25,28 +25,29 @@
     "@sentry/integrations": "^6.10.0",
     "@sentry/node": "^6.10.0",
     "@simplewebauthn/server": "^1.0.0",
-    "axios": "^0.19.0",
+    "axios": "^0.21.4",
     "axios-rate-limit": "^1.3.0",
     "axios-retry": "^3.1.8",
     "big.js": "^3.1.3",
     "body-parser": "^1.18.1",
     "compression": "^1.7.0",
     "cors": "^2.8.5",
-    "delay": "^4.4.0",
+    "delay": "^5.0.0",
     "ejs": "^2.6.1",
     "elliptic": "^6.5.3",
     "express": "^4.15.4",
-    "express-openapi-validator": "^4.12.7",
+    "express-openapi-validator": "^4.13.1",
     "helmet": "^3.21.0",
     "http-errors": "^1.8.0",
     "mongodb": "^3.1.10",
-    "p-forever": "^2.1.0",
+    "p-forever": "^3.0.1",
     "retry": "^0.10.1",
     "semver": "^7.3.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.13.16",
-    "@babel/preset-env": "^7.13.15",
+    "@babel/core": "^7.15.5",
+    "@babel/eslint-parser": "^7.15.4",
+    "@babel/preset-env": "^7.15.6",
     "@coinspace/b39": "^2.1.4",
     "@coinspace/cs-binance-smart-chain-wallet": "^1.0.0",
     "@coinspace/cs-eos-wallet": "^1.3.3",
@@ -77,8 +78,8 @@
     "dotenv": "^8.2.0",
     "dotenv-webpack": "^7.0.2",
     "ejs-loader": "^0.5.0",
-    "eslint": "^7.23.0",
-    "eslint-config-coinspace": "^1.1.3",
+    "eslint": "^7.32.0",
+    "eslint-config-coinspace": "^1.1.6",
     "events": "^3.3.0",
     "evp_bytestokey": "^1.0.3",
     "file-loader": "^6.2.0",
@@ -91,7 +92,7 @@
     "lodash": "^4.17.5",
     "mini-css-extract-plugin": "^1.5.1",
     "node-sass": "^4.5.3",
-    "nodemon": "^2.0.7",
+    "nodemon": "^2.0.12",
     "path-browserify": "^1.0.1",
     "postcss-loader": "^5.2.0",
     "qr.js": "0.0.0",
@@ -129,13 +130,14 @@
       "commonjs": true,
       "mocha": true
     },
+    "parser": "@babel/eslint-parser",
     "parserOptions": {
+      "requireConfigFile": false,
       "sourceType": "module"
     },
     "overrides": [
       {
         "files": [
-          "server/**",
           "phonegap/**",
           "cli/**",
           "electron/**"
