@@ -287,6 +287,6 @@ async function getAll() {
   ];
 }
 
-module.exports = {
+export default {
   getAll,
 };
