@@ -1,9 +1,6 @@
-'use strict';
-
-const Sentry = require('@sentry/node');
-const Integrations = require('@sentry/integrations');
-const db = require('./lib/v1/db');
-const tasks = require('./lib/tasks');
+import Sentry from '@sentry/node';
+import Integrations from '@sentry/integrations';
+import tasks from './lib/tasks.js';
 
 Sentry.init({
   dsn: process.env.SENTRY_DSN,
@@ -16,15 +13,13 @@ Sentry.init({
   ],
 });
 
-db().then(async () => {
-  await Promise.all([
-    tasks.syncTokens(12 * 60 * 60 * 1000), // delay 12 hours
-    tasks.updatePrices(60 * 1000), // delay 1 minute
-    tasks.cacheFees(5 * 60 * 1000), // 5 minutes
-    tasks.cacheMoonpayCurrencies(60 * 60 * 1000), // 1 hour
-    tasks.cacheMoonpayCountries(60 * 60 * 1000), // 1 hour
-    tasks.cacheGithubReleases(10 * 60 * 1000), // 10 minutes
-  ]);
-}).catch((error) => {
+await Promise.all([
+  tasks.syncTokens(12 * 60 * 60 * 1000), // delay 12 hours
+  tasks.updatePrices(60 * 1000), // delay 1 minute
+  tasks.cacheFees(5 * 60 * 1000), // 5 minutes
+  tasks.cacheMoonpayCurrencies(60 * 60 * 1000), // 1 hour
+  tasks.cacheMoonpayCountries(60 * 60 * 1000), // 1 hour
+  tasks.cacheGithubReleases(10 * 60 * 1000), // 10 minutes
+]).catch((error) => {
   console.error('error', error);
 });
