@@ -1,46 +1,44 @@
-'use strict';
+import createError from 'http-errors';
+import wallets from '../wallets.js';
+import storage from './storage.js';
+import mecto from './mecto.js';
+import moonpay from './moonpay.js';
+import tokens from '../tokens.js';
+import fee from '../fee.js';
+import csFee from '../csFee.js';
+import { verifyReq } from '../utils.js';
 
-const createError = require('http-errors');
-const wallets = require('./wallets');
-const storage = require('./storage');
-const mecto = require('./mecto');
-const moonpay = require('./moonpay');
-const tokens = require('../tokens');
-const fee = require('../fee');
-const csFee = require('../csFee');
-const { asyncWrapper, verifyReq } = require('./utils');
-
-exports.register = asyncWrapper(async (req, res) => {
+export async function register(req, res) {
   await verifyReq(req.body.walletId, req);
   const info = await wallets.register(req.body.walletId, req.body.deviceId, req.body.pinHash);
   console.log('registered wallet: %s device: %s', req.body.walletId, req.body.deviceId);
   res.status(201).send(info);
-});
+}
 
 // Public
 
-exports.tokenPublicPinVerify = asyncWrapper(async (req, res) => {
+export async function tokenPublicPinVerify(req, res) {
   await wallets.pinVerify(req.device, req.body.pinHash, 'public');
   return res.status(200).send({
     publicToken: req.device.public_token,
   });
-});
+}
 
-exports.tokenPublicPlatformOptions = asyncWrapper(async (req, res) => {
+export async function tokenPublicPlatformOptions(req, res) {
   const options = await wallets.platformOptions(req.device, 'public');
   return res.status(200).send(options);
-});
+}
 
-exports.tokenPublicPlatformVerify = asyncWrapper(async (req, res) => {
+export async function tokenPublicPlatformVerify(req, res) {
   await wallets.platformVerify(req.device, req.body, 'public');
   return res.status(200).send({
     publicToken: req.device.public_token,
   });
-});
+}
 
 // Private
 
-exports.tokenPrivate = asyncWrapper(async (req, res) => {
+export async function tokenPrivate(req, res) {
   if (req.device.wallet.settings['1fa_private'] === false) {
     if (req.device.wallet.authenticators.length === 0) {
       return res.status(200).send({
@@ -50,9 +48,9 @@ exports.tokenPrivate = asyncWrapper(async (req, res) => {
   }
 
   throw createError(401, 'Authorization required');
-});
+}
 
-exports.tokenPrivatePinVerify = asyncWrapper(async (req, res) => {
+export async function tokenPrivatePinVerify(req, res) {
   if (req.device.wallet.settings['1fa_private'] !== false) {
     await wallets.pinVerify(req.device, req.body.pinHash, 'private');
     if (req.device.wallet.authenticators.length === 0) {
@@ -66,18 +64,18 @@ exports.tokenPrivatePinVerify = asyncWrapper(async (req, res) => {
   }
 
   throw createError(400, 'Incorrect authenticator');
-});
+}
 
-exports.tokenPrivatePlatformOptions = asyncWrapper(async (req, res) => {
+export async function tokenPrivatePlatformOptions(req, res) {
   if (req.device.wallet.settings['1fa_private'] !== false) {
     const options = await wallets.platformOptions(req.device, 'private');
     return res.status(200).send(options);
   }
 
   throw createError(400, 'Incorrect authenticator');
-});
+}
 
-exports.tokenPrivatePlatformVerify = asyncWrapper(async (req, res) => {
+export async function tokenPrivatePlatformVerify(req, res) {
   if (req.device.wallet.settings['1fa_private'] !== false) {
     await wallets.platformVerify(req.device, req.body, 'private');
     if (req.device.wallet.authenticators.length === 0) {
@@ -91,71 +89,71 @@ exports.tokenPrivatePlatformVerify = asyncWrapper(async (req, res) => {
   }
 
   throw createError(400, 'Incorrect authenticator');
-});
+}
 
-exports.tokenPrivateCrossplatformOptions = asyncWrapper(async (req, res) => {
+export async function tokenPrivateCrossplatformOptions(req, res) {
   if (req.device.wallet.settings['1fa_private'] === false) {
     const options = await wallets.crossplatformOptions(req.device, 'private');
     return res.status(200).send(options);
   }
 
   throw createError(400, 'Incorrect authenticator');
-});
+}
 
-exports.tokenPrivateCrossplatformVerify = asyncWrapper(async (req, res) => {
+export async function tokenPrivateCrossplatformVerify(req, res) {
   await wallets.crossplatformVerify(req.device, req.body, 'private');
   return res.status(200).send({
     privateToken: req.device.private_token,
   });
-});
+}
 
 // Attestation
 
-exports.platformAttestationOptions = asyncWrapper(async (req, res) => {
+export async function platformAttestationOptions(req, res) {
   const options = await wallets.platformAttestationOptions(req.device);
   res.status(200).send(options);
-});
+}
 
-exports.platformAttestationVerify = asyncWrapper(async (req, res) => {
+export async function platformAttestationVerify(req, res) {
   await wallets.platformAttestationVerify(req.device, req.body);
   res.status(200).send({ success: true });
-});
+}
 
-exports.crossplatformAttestationOptions = asyncWrapper(async (req, res) => {
+export async function crossplatformAttestationOptions(req, res) {
   const options = await wallets.crossplatformAttestationOptions(req.device);
   res.status(200).send(options);
-});
+}
 
-exports.crossplatformAttestationVerify = asyncWrapper(async (req, res) => {
+export async function crossplatformAttestationVerify(req, res) {
   await wallets.crossplatformAttestationVerify(req.device, req.body);
   res.status(200).send({ success: true });
-});
+}
 
 // API
 
-exports.removePlatformAuthenticator = asyncWrapper(async (req, res) => {
+export async function removePlatformAuthenticator(req, res) {
   await wallets.removePlatformAuthenticator(req.device);
   res.status(200).send({ success: true });
-});
+}
 
-exports.listCrossplatformAuthenticators = asyncWrapper(async (req, res) => {
+export async function listCrossplatformAuthenticators(req, res) {
   const list = await wallets.listCrossplatformAuthenticators(req.device);
   res.status(200).send(list);
-});
+}
 
-exports.removeCrossplatformAuthenticator = asyncWrapper(async (req, res) => {
+export async function removeCrossplatformAuthenticator(req, res) {
   await wallets.removeCrossplatformAuthenticator(req.device, req.body.credentialID);
   res.status(200).send({ success: true });
-});
+}
 
-exports.getSettings = asyncWrapper(async (req, res) => {
+export async function getSettings(req, res) {
   res.status(200).send({
     '1faPrivate': req.device.wallet.settings['1fa_private'],
     hasAuthenticators: req.device.wallet.authenticators.length !== 0,
   });
-});
+}
 
-exports.setSettings = asyncWrapper(async (req, res) => {
+export async function setSettings(req, res) {
   const data = {};
   if ('1faPrivate' in req.body) {
     data['1fa_private'] = req.body['1faPrivate'];
@@ -165,89 +163,89 @@ exports.setSettings = asyncWrapper(async (req, res) => {
     '1faPrivate': settings['1fa_private'],
     hasAuthenticators: req.device.wallet.authenticators.length !== 0,
   });
-});
+}
 
-exports.getDetails = asyncWrapper(async (req, res) => {
+export async function getDetails(req, res) {
   res.status(200).send({
     data: req.device.wallet.details,
   });
-});
+}
 
-exports.setDetails = asyncWrapper(async (req, res) => {
+export async function setDetails(req, res) {
   const data = await wallets.setDetails(req.device, req.body.data);
   res.status(200).send({ data });
-});
+}
 
-exports.getStorage = asyncWrapper(async (req, res) => {
+export async function getStorage(req, res) {
   const data = await storage.getStorage(req.device, req.params.storageName);
   res.status(200).send({ data });
-});
+}
 
-exports.setStorage = asyncWrapper(async (req, res) => {
+export async function setStorage(req, res) {
   const data = await storage.setStorage(req.device, req.params.storageName, req.body.data);
   res.status(200).send({ data });
-});
+}
 
-exports.setUsername = asyncWrapper(async (req, res) => {
+export async function setUsername(req, res) {
   const username = await wallets.setUsername(req.device, req.body.username);
   res.status(200).send({ username });
-});
+}
 
-exports.removeDevice = asyncWrapper(async (req, res) => {
+export async function removeDevice(req, res) {
   await wallets.removeDevice(req.device);
   res.status(200).send({ success: true });
-});
+}
 
-exports.removeWallet = asyncWrapper(async (req, res) => {
+export async function removeWallet(req, res) {
   await wallets.removeWallet(req.device);
   res.status(200).send({ success: true });
-});
+}
 
-exports.searchMecto = asyncWrapper(async (req, res) => {
+export async function searchMecto(req, res) {
   const results = await mecto.search(req.device, req.query);
   res.status(200).send(results);
-});
+}
 
-exports.saveMecto = asyncWrapper(async (req, res) => {
+export async function saveMecto(req, res) {
   await mecto.save(req.device, req.body);
   res.status(200).send({ success: true });
-});
+}
 
-exports.removeMecto = asyncWrapper(async (req, res) => {
+export async function removeMecto(req, res) {
   await mecto.remove(req.device);
   res.status(200).send({ success: true });
-});
+}
 
-exports.moonpaySign = asyncWrapper(async (req, res) => {
+export async function moonpaySign(req, res) {
   const urls = moonpay.sign(req.body.urls);
   res.status(200).send({ urls });
-});
+}
 
-exports.getTokens = asyncWrapper(async (req, res) => {
+export async function getTokens(req, res) {
   const list = await tokens.getTokens(req.query.network);
   res.status(200).send(list);
-});
+}
 
-exports.getTicker = asyncWrapper(async (req, res) => {
+export async function getTicker(req, res) {
   const ticker = await tokens.getTicker(req.query.crypto);
   res.status(200).send({
     _id: ticker._id,
     prices: ticker.prices,
     // strip decimals
   });
-});
+}
 
-exports.getTickers = asyncWrapper(async (req, res) => {
+export async function getTickers(req, res) {
   const tickers = await tokens.getTickers(req.query.crypto);
   res.status(200).send(tickers);
-});
+}
 
-exports.getFees = asyncWrapper(async (req, res) => {
+export async function getFees(req, res) {
   const fees = await fee.getFees(req.query.crypto);
   res.status(200).send(fees);
-});
+}
 
-exports.getCsFee = asyncWrapper(async (req, res) => {
+export async function getCsFee(req, res) {
   const fee = await csFee.getCsFee(req.query.crypto);
   res.status(200).send(fee);
-});
+}
