@@ -1,20 +1,17 @@
-'use strict';
+import tokens from '../tokens.js';
+import cryptos from '../cryptos.js';
 
-const tokens = require('../tokens');
-const cryptos = require('../cryptos');
-const { asyncWrapper } = require('../v2/utils');
-
-exports.qwe = asyncWrapper(async (req, res) => {
+export async function qwe(req, res) {
   return res.status(200).send('ok');
-});
+}
 
-exports.getCryptos = asyncWrapper(async (req, res) => {
+export async function getCryptos(req, res) {
   const list = await cryptos.getAll();
   res.status(200).send(list);
-});
+}
 
-exports.getTickers = asyncWrapper(async (req, res) => {
+export async function getTickers(req, res) {
   // TODO: move to crypto collection
   const tickers = await tokens.getTickersV3(req.query.crypto);
   res.status(200).send(tickers);
-});
+}
